fixed Serological vs PCR text (written wrong in the slideshow)
hardcoded properly in the smart contract, said wrong in the video presentation.

To donate plasma a person must have tested POSITIVE on the Serological Test for Covid 19 (showing that they have antibodies) and NEGATIVE on the PCR Test (confirming that they don't have Covid19 currently).

PDF File was wrong. YouTube video is also wrong.
I'm sorry for the confusion. The Smart Contract and App demonstrate the proper logic, though...